backend/requirements.txt
Django>=4.2
djangorestframework
Pillow
qrcode
psycopg2-binary   # if using Postgres
djangorestframework-simplejwt  # for JWT auth
django-cors-headers
